+++ openconnect-internal.h
@@ -455,7 +455,7 @@ int script_config_tun(struct openconnect_info *vpninfo, const char *reason);
 void os_shutdown_tun(struct openconnect_info *vpninfo);
 int os_read_tun(struct openconnect_info *vpninfo, struct pkt *pkt, int new_pkt);
 int os_write_tun(struct openconnect_info *vpninfo, struct pkt *pkt);
-int os_setup_tun(struct openconnect_info *vpninfo);
+intptr_t os_setup_tun(struct openconnect_info *vpninfo);
 
 /* dtls.c */
 unsigned char unhex(const char *data);
